requirements.txt
python-fasthtml
passlib
litellm>=1.12.0
python-dotenv
bcrypt
email-validator
python-multipart